% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_length_summary.R
\name{average_seq_length}
\alias{average_seq_length}
\title{Calculate the Average Length of Sequences}
\usage{
average_seq_length(sequences)
}
\arguments{
\item{sequences}{A character vector where each element is a sequence of elements separated by spaces.}
}
\value{
A numeric value representing the average length of the sequences.
}
\description{
This function calculates the average length of a sequence of elements, where each sequence is split by spaces.
}
\examples{
sequences <- c('hello world', 'hello world hello', 'hello world hello world')
average_seq_length(sequences)
}
