% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_conditional_entropy.R
\name{calculate_conditional_entropy}
\alias{calculate_conditional_entropy}
\title{Calculate Conditional Entropy of B given A in Bits}
\usage{
calculate_conditional_entropy(vectorA, vectorB)
}
\arguments{
\item{vectorA}{A categorical vector representing the conditioning variable A}

\item{vectorB}{A categorical vector representing the conditioned variable B}
}
\value{
The conditional entropy of B given A in bits
}
\description{
This function calculates the conditional entropy of B given A in bits between two categorical vectors
}
\examples{
vectorA <- c("A", "B", "A", "A", "B", "C", "C", "C", "A", "B")
vectorB <- c("X", "Y", "X", "Y", "X", "Y", "Y", "X", "X", "Y")
calculate_conditional_entropy(vectorA, vectorB)
}
