% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_length_summary.R
\name{median_seq_length}
\alias{median_seq_length}
\title{Calculate the Median Length of Sequences}
\usage{
median_seq_length(sequences)
}
\arguments{
\item{sequences}{A character vector where each element is a sequence of elements separated by spaces.}
}
\value{
A numeric value representing the median length of the sequences.
}
\description{
This function calculates the median length of a sequence of elements, where each sequence is split by spaces.
}
\examples{
sequences <- c('hello world', 'hello world hello', 'hello world hello world')
median_seq_length(sequences)
}
