% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_length_summary.R
\name{min_max_seq_length}
\alias{min_max_seq_length}
\title{Calculate the Minimum and Maximum Length of Sequences}
\usage{
min_max_seq_length(sequences)
}
\arguments{
\item{sequences}{A character vector where each element is a sequence of elements separated by spaces.}
}
\value{
A numeric vector of length 2, with the minimum and maximum lengths of the sequences.
}
\description{
This function calculates the minimum and maximum length of sequences of elements, where each sequence is split by spaces.
}
\examples{
sequences <- c('hello world', 'hello world hello', 'hello world hello world')
min_max_seq_length(sequences)
}
