% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sequence_length_distributions.R
\name{plot_seq_length_distribution}
\alias{plot_seq_length_distribution}
\title{Plot the Distribution of Sequence Lengths}
\usage{
plot_seq_length_distribution(sequences)
}
\arguments{
\item{sequences}{A character vector where each element is a sequence of elements separated by spaces.}
}
\value{
A `ggplot` object showing the distribution of sequence lengths.
}
\description{
This function plots the distribution of the lengths of sequences of elements, where each sequence is split by spaces. The plot includes a histogram and a vertical line indicating the mean length.
}
\examples{
sequences <- c('hello world', 'hello world hello', 'hello world hello world')
plot_seq_length_distribution(sequences)
}
