% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_length_summary.R
\name{sd_seq_length}
\alias{sd_seq_length}
\title{Calculate the Standard Deviation of Sequence Lengths}
\usage{
sd_seq_length(sequences)
}
\arguments{
\item{sequences}{A character vector where each element is a sequence of elements separated by spaces.}
}
\value{
A numeric value representing the standard deviation of the lengths of the sequences.
}
\description{
This function calculates the standard deviation of the lengths of sequences of elements, where each sequence is split by spaces.
}
\examples{
sequences <- c('hello world', 'hello world hello', 'hello world hello world')
sd_seq_length(sequences)
}
